Refactoring of Freight method.
- Refactoring of ServicePoint and TimeTable methods
- Use logrus instead of fmt.print
- Test data udpate
