--- freight/servicepoint/models/nearest_service_points_response.go
@@ -1,6 +1,0 @@
-package models
-
-type GetNearestServicePointsResponse struct {
-	XMLName       string `xml:"http://DHL.ServicePoint.ServiceContracts/2008/10 GetNearestServicePointsResponse"`
-	ServicePoints *ServicePoints
-}
