--- freight/servicepoint/models/service_point_details_request.go
@@ -1,6 +1,0 @@
-package models
-
-type GetServicePointDetailRequest struct {
-	XMLName         string           `xml:"http://DHL.ServicePoint.ServiceContracts/2008/10 GetServicePointDetailRequest"`
-	ServicePointRef *ServicePointRef `xml:"http://DHL.ServicePoint.DataContracts/2008/10 ServicePointRef"`
-}
