--- freight/servicepoint/models/service_point_ref.go
@@ -1,6 +1,0 @@
-package models
-
-type ServicePointRef struct {
-	ID          string `xml:"Id,omitempty"`
-	DisplayName string `xml:"DisplayName,omitempty"`
-}
