--- freight/servicepoint/models/service_points.go
@@ -1,6 +1,0 @@
-package models
-
-type ServicePoints struct {
-	XMLName            string               `xml:"ServicePoints"`
-	NearbyServicePoint []NearbyServicePoint `xml:"http://DHL.ServicePoint.DataContracts/2008/10 NearbyServicePoint"`
-}
